fixed errors

+++ project_4/Project_4/source/test.c
@@ -17,20 +17,20 @@ void Test(void) {
 	UCUNIT_Init(); /* initialize framework */
 	UCUNIT_TestcaseBegin("State_Machine_Transitions");
 	HandleEventComplete();
-	UCUNIT_CheckIsEqual(test_event, kProcessDisplay);
+	UCUNIT_CheckIsEqual((int)test_event, (int)kProcessDisplay);
 	HandleEventComplete();
-	UCUNIT_CheckIsEqual(test_event, kWaitPollSlider);
+	UCUNIT_CheckIsEqual((int)test_event, (int)kWaitPollSlider);
 	HandleEventTimeout_1_5();
-	UCUNIT_CheckIsEqual(test_event, kReadXYZ);
+	UCUNIT_CheckIsEqual((int)test_event, (int)kReadXYZ);
 
 	test_event=kWaitPollSlider;
 	HandleEventTimeout_6();
-	UCUNIT_CheckIsEqual(test_event, kReadXYZ);
-	UCUNIT_CheckIsEqual(test_machine, mTableDriven);
+	UCUNIT_CheckIsEqual((int)test_event, (int)kReadXYZ);
+	UCUNIT_CheckIsEqual((int)test_machine, (int)mTableDriven);
 	test_event=kWaitPollSlider;
 	HandleEventLeftSlider();
-	UCUNIT_CheckIsEqual(test_event, kReadXYZ);
-	UCUNIT_CheckIsEqual(test_machine, mStateCentric);
+	UCUNIT_CheckIsEqual((int)test_event, (int)kReadXYZ);
+	UCUNIT_CheckIsEqual((int)test_machine, (int)mStateCentric);
 
 	UCUNIT_TestcaseEnd();
